Refactor: Move some indivudal components from ItemDetailsList to single one
